+++ src/admin/pages/ContentType/Edit/CreateField/fieldTypes/CreateChoice/index.tsx
