+++ src/admin/pages/ContentType/Edit/CreateField/fieldTypes/FileImage/index.tsx
